nearly complete with widgets, just one kind of fundamental issue to work around

+++ ReadySetWidgets/SimpleEntry.swift
@@ -10,6 +10,6 @@ import WidgetKit
 
 struct SimpleEntry: TimelineEntry {
     let date: Date
-    let consumption: Double
+    let consumption: Int
     let goal: Double
 }
